[master] scudo: Re-order Allocator fields for improved performance. NFCI.
Move smaller and frequently-accessed fields near the beginning
of the data structure in order to improve locality and reduce
the number of instructions required to form an access to those
fields. With this change I measured a ~5% performance improvement on
BM_malloc_sql_trace_default on aarch64 Android devices (Pixel 4 and
DragonBoard 845c).

Differential Revision: https://reviews.llvm.org/D88350

GitOrigin-RevId: e851aeb0a5084d968d6384fbc2257bbe05dcdacb
Change-Id: I642559723891a34cd281e5b0b5bd9a28d274129c
Signed-off-by: Jake Weinstein <jake@aospa.co>
Signed-off-by: AshutoshSundresh <ashutoshsundresh@gmail.com>